requirements.txt
numpy<2
pandas
requests
beautifulsoup4
nltk
textblob
textstat
openpyxl
streamlit
plotly
scikit-learn
spacy
yake
wordcloud
reportlab
kaleido
rake_nltk
newspaper3k
lxml[html_clean]
readability-lxml
lxml